+++ src/components/views/to-dos/header/index.tsx
@@ -1,4 +1,4 @@
-import { useDates } from '../../../../../hooks/useDates'
+import { useDates } from '../../../../hooks/useDates';
 
 export default function Header() {
     const { dates } = useDates()
